chore: rework of types
